Decouple the Jenkins slave name from AWS instance
This change is meant to decouple the Jenkins slave's name that is used
in the "openzfs-jenkins-slave" ansible role, from the AWS instance id of
the slave. Having the tight coupling of the slave name and the instance
id means that it's not possible to re-use that role for a system that is
not running on AWS. Now, the slave's name must be a user specified value
that is passed into the role via that "jenkins_name" variable.

As you might expect, the existing Jenkins jobs that used this role also
had to be updated. The "create-build-slave" job was updated to take a
new "SLAVE_NAME" parameter as input. Also, the "openzfs-precommit" job
was updated to pass in its "$BUILD_TAG" variable to use for the name of
the slave it creates. Thus, now when an "openzfs-precommit" job is
kicked off, new slaves will connect and have more meaningful names,
since the name will reference back to the job that spawned it (as
opposed to reference the slave's AWS instance id).

And finally, the "AWS_INSTANCE_ID" parameter was removed from the
"openzfs-build-nightly" job since it isn't actually used anywhere.