Add "GIT_BRANCH" parameter to "build-nightly" job
This change adds a new parameter to the "openzfs-build-nightly" job to
allow a consumer to provide a specific branch or git hash that should be
checked out, built, and tested. This allows the job to be more flexible,
and is in anticipation of using this in conjunction with building pull
requests; in which case the pull request's git hash will need to be
specified.