Remove the usage of the "ansible_sudo" directive
When using the "ansible_sudo" directive and the "become" directive at
the same time, the "become" directive becomes ineffective. Within the
"openzfs-jenkins-slave" role, we need to run some tasks as the root
user, and some tasks as a non-root user. Thus, to accomplish both, we've
removed the host global "ansible_sudo" directive, and replaced it with
task specific "become" and "become_user" directives.

Additionally, since the "ansible_sudo" usage isn't used anymore, the
platform specific "ssh_sudo" variable has been removed from each of the
platforms' credentials file and removed the usage of this variable.