Rework and move nightly_env_set_var function
This patch modifies the "nightly_env_set_var" function, with the
intention of making it more understandable given the limited
functionality that is needed from it.

The optional "-f" parameter was removed. This was the only way in which
the command was being called, so now this is the default and only way
the function behaves. It didn't make sense to support the "non force"
use case since that wasn't being exercised.