Add ability to only unregister DCenter instances
This change adds the ability to only unregister a DCenter instance after
a job is completed, instead of completely destroying the instance. This
functionality is useful when debugging failures, as it's often necessary
to inspect the instance where the failure occurred, in order to root
cause the problem.

Currently the parameter isn't being used for anything, but the plan is
to incorporate this into the high level "precommit" job such that on
failure, the instance is unregistered instead of destroyed. This way the
instance's state is preserved for later analysis.