Use shorter slave name to "fix" connection issues
The root cause is still unknown, but to workaround some SSH connection
issues experienced when Ansible attempts to connect to the Jenkin's
slave, the slave's name has been shortened. When using the $BUILD_TAG as
the Jenkins slave's name, Ansible will fail to connect to the new slave
in the context of the "create-dc-build-slave" job reporting the error:

    fatal: [ ... ] => SSH Error: Permission denied ...
        while connecting to 10.43.4.88:22

On a whim, a shorter name was given to the slave as a quick test, and,
lo and behold, Ansible was able to connect to the newly created slave.
Thus, hopefully as a temporary workaround, a new convention has been
adopted for assigning names to the DCenter build slaves, which is short
enough to keep Ansible happy.