Add "deploy_role.py" script for easier development
This patch adds a new "deploy_role.py" utility which is intended to make
development and testing of new and/or existing roles easier. Instead of
having to create new inventory files and playbooks in order to deploy a
role to a host in a one-off fashion, this utility can be used.

As an example, the following command would run the "initialize-omnios"
role on the host, "psurya-omnios.dcenter", using the specified
credentials:

    $ python bin/deploy_role.py -H psurya-omnios.dcenter \
          -r initialize-omnios -u root -p password