Remove AWS specific Jenkins jobs
This patch removes the OpenZFS Jenkins jobs specific for running on top
of Amazon EC2. Testing has shown that illumos does not run on AWS, so we
can't yet use that platform to run the regression tests. Thus, it
doesn't make sense to keep these Jenkins jobs in the tree; they'll
remain in the history, though, if they need to be referenced at a later
time (e.g. if AWS support lands on illumos).