Rename remaining Jenkins jobs
This is a simple change that is not intended to change functionality,
but only to serve as some minor cleanup of the Jenkins job names now
that the Amazon specific jobs have been removed. Apart from some minor
fixes to comments within the jobs, the names and file names of the jobs
have changed:

    * "create-dc-build-slave" was renamed to "create-build-slave"
    * "destroy-dc-build-slave" was renamed to "destroy-build-slave"
    * "openzfs-precommit-dc" was renamed to "openzfs-regression-tests"

Along with the job renames, the files were renamed to reflect the name
change as well.