Remove Jenkins variables from build scripts
This small patches removes the usage of Jenkins provided environment
variables from the "build-os.sh" and "install-os.sh" scripts. This was
done to make them easier to run outside of a Jenkins job (e.g. when
using "deploy_role.sh" during development), as well as fix a bug when
running the "install-os.sh" script within the Jenkins job (the BUILD_TAG
variable isn't defined within that script, since the environment is
cleared when running sudo).