Clone the DCenter instance after a build completes
This change adds support to clone the DCenter instances that does the
build of OpenZFS into two new Jenkins slaves. The idea, is that after a
build is complete, that instance will have installed the new build
products into its root filesystem.

From this system, we can clone the instance into two new slaves and run
the OpenZFS test jobs in parallel on these two new slaves. Since the
base system that was cloned had the build products installed to its root
filesystem prior to the clone operation, each new cloned slave should be
running the new build products when they boot up.

In order to add this support, a new Ansible role had to be created,
"clone-dc-instance"; as well as the Ansible playbook that uses this new
role, "clone-build-slave". This playbook is called from the new Jenkins
job, "clone-build-slave", and the "openzfs-regression-tests" was updated
to execute this new job after the OpenZFS build completes. And finally,
the "destroy-build-slave" Jenkins job had to be updated to destroy the
original DCenter instance, as well as the two new clones that are
created.