Split key, value strings in properties file tasks
The main change here is we properly split the "key" and the "value"
parts of the items when using "lineinfile" when writing out the
properties files. This is needed to ensure we overwrite any existing
values for the keys, otherwise we could potentially write a new line
without removing the old value.

Additionally, the instance's DNS information is no longer being written
out, as that information is no longer used.