Add OpenZFS functional test suite to parent job
This simple change adds the OpenZFS functional regression test suite
into the mix, such that it runs in parallel to the OpenZFS ztest run; it
will run on one of the cloned build slaves. To do this, a new script was
added to the build slave, and the new Jenkins job to run this script was
created.