Configure build slave clones individually
For some odd reason (that I don't understand), when assigning the build
clones the same Ansible group and configuring them concurrently, one of
the systems will fail to connect resulting in the whole job to fail.

As a workaround, each clone is assigned to a unique Ansible group, and
each slave is configured serially and in isolation from the other clone.
This appears to cause the "clone-build-slave" job to be more stable,
although the reason for the change in stability due to this change is
still not understood.