Revert back to "copy" module in create-build-slave
Since we're no longer using the DC_INSTANCE_DNS variable, we can safely
revert back to using the "copy" module in the "create-build-slave.yml"
Ansible playbook instead of the "lineinfile" module.

The reason for doing this, is we need to truncate the properties file
for each new regression test run, in order to clear out any prior job's
instance names that might still be lingering in the file. In particular,
if the DC_INSTANCE_NAME_A or DC_INSTANCE_NAME_B variables are present
from a past run, and the "create-build-slave" job fails, then the
"destroy-build-slave" job will be passed in the old instance names.
Clearing out the file, using the "copy" module, prevents this issue.