Remove "unregister-dc-instance" Ansible role
This patch removes the "unregister-dc-instance" role, and instead
adds a conditional to the "destroy-dc-instance" role that can be
used to optionally unregister the slave instead of destroying it.

Additionally, the "destroy-dc-instance" role was updated to not fail if
the "instance_name" variable isn't defined. Instead, this role will skip
the destroy or unregister attempt if this is the case. This is needed,
because it's possible for the "create-build-slave" job to fail, in which
case "destroy-build-slave" will be run without cloned instance names to
fill in all of the fields of the destroy job.