Add "killPhaseCondition" to regression tests
This patch adds the "killPhaseCondition" parameter to the regression
test jobs that are run. This should ensure each job is run until
completion, even if one of the other jobs in the phase fails. Otherwise,
all jobs would be killed after the first failed job in the phase.
Additionally, the clone and regression tests were enabled to allow these
new parameter to be tested.