Remove Vagrant since it's not used and out of date
Using Vagrant to speed up development of the Jenkins master instance was
initially a good idea, but it hasn't been used lately and doesn't
provide much benefit since the Jenkins jobs can't be executed from the
Vagrant environment. This commit removes the file; it doesn't make sense
to maintain it moving forward since it's not being used regularly.