Disable specific Jenkins jobs on the public mirror
This patch updates the Ansible playbook that synchronizes the public
Jenkins instance, to also disable specific Jenkins jobs on the public
instance that would otherwise cause issues. For example, if the
"openzfs-regression-tests" job is not disabled on the public instance,
this job would attempt to poll the GitHub project page and compete with
the polling done by the internal Jenkins instance.

Closes #12