Add `-l` to the `NIGHTLY_OPTIONS` of the build
This patch adds support for running lint during the build by fixing the
`SPRO_ROOT` and `ONLY_LINT_DEF` variables for running on OmniOS.
Previously I was using `/opt` for `SPRO_ROOT` and basing `ONLY_LINT_DEF`
off of the `/opt/SUNWspro` directory. Both of these values are incorrect
for OmniOS, and instead both should be using the `/opt/sunstudio12.1`
directory; this patch makes that changes and enables the `-l` option in
`NIGHTLY_OPTIONS`.

Closes #2