Add script to parse the ZFS test suite results
This patch is intended to make the results of the "openzfs-run-zfs-test"
easier to consume. Currently the "openzfs-run-zfs-test" job will always
report success, even if there are failed tests within the ZFS test suite
run. This makes it a very involved process to determine if a given run
has uncovered any new failures or not.

This patch attempts to address this by adding a script that will
post-process the results of all the individual ZFS tests, and report an
error if there are new failed tests that cropped up during any given
run. In addition, since there are a few known tests to fail, these
specific failures will not cause the "openzfs-run-zfs-test" job to fail.
Instead, these specific failures are reported but exempt from causing
the Jenkins job to fail, and any unexpected results (e.g. unexpected
failures, or passing tests that are expected to fail) will result in the
Jenkins job to fail.

Additionally, the script that parses the results will also print a more
compact summary which is intended to be consumed by humans. Here's an
example of the scripts output:

    $ python zfstest-report.py results.txt

    Tests with results other than PASS that are expected:
        FAIL cli_root/zfs_property/zfs_written_property_001_pos
        FAIL cli_root/zfs_mount/zfs_mount_all_001_pos
        FAIL mdb/mdb_001_pos
        FAIL rsend/rsend_009_pos
        FAIL zvol/zvol_swap/zvol_swap_004_pos
        FAIL rootpool/rootpool_002_neg
        FAIL acl/nontrivial/zfs_acl_chmod_inherit_003_pos
        FAIL cli_root/zfs_mount/zfs_mount_009_neg
        FAIL grow_replicas/grow_replicas_001_pos
        FAIL rsend/rsend_008_pos
        FAIL cli_root/zfs_snapshot/zfs_snapshot_009_pos
        FAIL refreserv/refreserv_004_pos
        FAIL cli_root/zpool_get/zpool_get_002_pos

    Tests with results other than PASS that are unexpected:

    Tests with result of PASS that are unexpected:

    Percent passed: 98.53%
    Log directory:  /var/tmp/test_results/20151025T072352