Don't fail if tests unexpectedly PASS
The important change in this patch is the fact that tests that
unexpectedly PASS no longer fail the test as a whole. These are
reported, just like expected FAIL, but they no longer will cause the
script to return a non-zero exit status.

Additionally, the `zvol/zvol_misc/zvol_misc_002_pos` test was added to
the list of expected failures.

And finally, the list of expected failures was rearranged such that it
is sorted alphabetically. This should make it more clear where to insert
a new test in this list, if another addition is needed in the future.