OpenZFS Regression Tests Version 2 using Workflow
This patch makes a first pass attempt at implementing the current
"openzfs-regression-tests" job using the Jenkins Workflow plugin. This
new job is not yet configured to automatically trigger based off of
incoming pull requests, but allows a user to manually trigger the job
and pass in a branch or SHA hash to build and test. This allows the new
job to be tested, and get some soak time, before making any decisions
about if it should be used by default of OpenZFS regression tests.

Additionally, this patch makes a few additional changes:

    * The Jenkins swarm slaves now use the same slave name as their
      DCenter instance name. This makes it easier to map a given Jenkins
      job, that's running on a slave, back to the DCenter instance that
      is responsible for the job; this makes manual debugging of
      failures easier.

    * The jobs interacting with the upstream "openzfs-build" repository
      have been updated to allow the owner and branch to be specified as
      parameters. This makes it easier to develop and test changes, and
      there's a single place to update in order to force these jobs to
      use the in-progress version of the repository.

    * The "destroy-build-slave" was updated to be tolerant of cases
      where any of the DC_INSTANCE_NAME variables might not be present,
      utilizing the "conditional-buildstep" plugin.