Add vsftpd and anonymous FTP downloads
This change installs the 'vsftpd' package to the Jenkins master
instances and configures them to allow for anonymous login and
downloads. This is useful when build and/or test results are needed by
external contributors, but the results are too big to be uploaded
directory to GitHub (e.g. ztest vdev file tarballs can be +50MB, and
GitHub limits comment attachments to 10MB). Ideally, we'd allow
contributors direct access to the VMs used for testing, but we can't
provide that until we move off of using Delphix's internal cloud.