Use data disk on public Jenkins mirror
The public Jenkins mirror is currently using Joyent's public cloud and
the system has a 64G disk attached to it, in addition to the disk used
for the root filesystem. Prior to this change, we were using the root
filesystem to store the Jenkins files, and since the disk used for the
root filesystem is only 8G, we managed to fill the root filesystem. This
patch configures the public mirror to use the 64G disk to store the
Jenkins files, the "/dev/vdb" device.