Bump the DCenter guest expiration to 7 days
Occasionally a patch will go through the test suite, fail, and the
DCenter guest will be cleaned up before a developer has the chance to
notice and investigate the failure. Thus, I'm bumping the expiration
to 7 days; this will hopefully be enough time, such that any failures
can be investigated prior to the DCenter guest being destroyed.