Block IP address range(s) using ufw
Joyent's support folks informed us of higher than normal DNS activity
coming from our public Jenkins master. Upon further investigation, a
rogue process was discovered that was continually sending network
traffic to the following IP address: 118.193.198.16

Using a combination of "tcpdump" and "lsof" the process responsible was
found to be something with the name of "xudp_udp" as can be seen by the
following snippet from "lsof":

    $ sudo lsof -i :44453
    COMMAND    PID    USER   FD   TYPE  DEVICE SIZE/OFF NODE NAME
    xudp_udp 17084 jenkins    4u  IPv4 5262887      0t0  TCP 33087b18-1709-475a-fb24-d35ff910ee9e:44453->118.193.198.16:5888 (ESTABLISHED)

Digging in some more, it became apparent that this process had
downloaded a bunch of scripts and executables into the "/tmp" directory
on that server:

    $ sudo ls /proc/17084/cwd
    2huizhen                      jna--1712433994  weiwei113.pl.6  weiwei118.pl    weiwei201621.pl
    2.pl                          sishen           weiwei113.pl.7  weiwei118.pl.1  weiwei201621.pl.1
    back.pl                       sshd             weiwei113.pl.8  weiwei118.pl.2  weiwei21.pl
    cmd.n                         tf1450           weiwei113.pl.9  weiwei118.pl.3  weiwei3-3.pl
    conf.n                        topk             weiwei117.pl    weiwei118.pl.4  weiwei35.pl
    etc                           udp              weiwei117.pl.1  weiwei118.pl.5  weiwei35.pl.1
    fedoIckm                      uhuizhen         weiwei117.pl.2  weiwei18.pl     weiwei456.pl
    fre25000                      vga.conf         weiwei117.pl.3  weiwei19.pl     weiwei.pl
    gates.lod                     weiwei113.pl     weiwei117.pl.4  weiwei19.pl.1   weiwei.pl.1
    hsperfdata_jenkins            weiwei113.pl.1   weiwei117.pl.5  weiwei19.pl.2   weiwei.pl.2
    hsperfdata_root               weiwei113.pl.2   weiwei117.pl.6  weiwei19.pl.3   weiwei.pl.3
    huizhen                       weiwei113.pl.3   weiwei117.pl.7  weiwei19.pl.4   winstone6744553430951277080.jar
    is                            weiwei113.pl.4   weiwei117.pl.8  weiwei19.pl.5   xudp_udp
    jetty-0.0.0.0-8080-war--any-  weiwei113.pl.5   weiwei117.pl.9  weiwei19.pl.6

That process was killed, the files and directories in "/tmp" removed,
and the system was rebooted; so far it looks as though the rogue network
traffic has ceased.

To try and prevent this from happening again, "whois" was used to
determine the IP address range that the destination address belonged to,
and this range of IP address have been blocked using "ufw".

Doing some quick Google searches, it looks like allowing passwordless
access to Jenkins is a Bad Idea, since Jenkins allows for arbitrary
groovy code to be run on the system and arbirary system commands can be
run as a result (as the user running the Jenkins server, i.e.
"jenkins"). Since this public instance doesn't have any "important"
information stored on it, this isn't a big deal, but we should try to
address this when possible.

So far, this has just been more of an inconvenience, than anything.